chore(Visibility): move logic to hook (#3698)
So we can use it in other contexts if needed.
